GPIO modification for RA4L1

+++ targets/arm/mikroe/renesas/src/gpio/hal_ll_gpio.c
@@ -61,7 +61,7 @@ void hal_ll_gpio_configure_pin(hal_ll_gpio_pin_t *pin, hal_ll_pin_name_t name, h
  */
 #if (FLATTEN_ME_LEVEL < FLATTEN_ME_LEVEL_LOW)
 uint8_t hal_ll_gpio_read_pin_input(hal_ll_gpio_pin_t *pin) {
-    uint8_t gpio_data_value;
+    uint16_t gpio_data_value;
 
     gpio_data_value = ((hal_ll_gpio_base_handle_t *)(pin->base))->pidr;
     return (gpio_data_value & pin->mask) ? 0x01 : 0x00;
@@ -73,9 +73,9 @@ uint8_t hal_ll_gpio_read_pin_input(hal_ll_gpio_pin_t *pin) {
  */
 #if (FLATTEN_ME_LEVEL < FLATTEN_ME_LEVEL_LOW)
 uint8_t hal_ll_gpio_read_pin_output(hal_ll_gpio_pin_t *pin) {
-    uint8_t gpio_data_value;
+    uint16_t gpio_data_value;
 
-    gpio_data_value = ((hal_ll_gpio_base_handle_t *)(pin->base))->podr;
+    gpio_data_value = ((hal_ll_gpio_base_handle_t *)(pin->base))->podr; 
     return (gpio_data_value & pin->mask) ? 0x01 : 0x00;
 }
 #endif
